requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Vector store
chromadb>=0.4.0

# Document processing
pypdf2>=3.0.0
python-docx>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Additional utilities
numpy>=1.24.0
pandas>=2.0.0